# Conflicts:
#	bindings/Python/sklearn_like/svc.cpp
#	include/plssvm/backends/CUDA/kernel/cg_explicit/blas.cuh
#	include/plssvm/backends/CUDA/kernel/cg_implicit/kernel_matrix_assembly_blas.cuh
#	include/plssvm/backends/CUDA/kernel/predict_kernel.cuh
#	include/plssvm/backends/HIP/kernel/cg_explicit/blas.hip.hpp
#	include/plssvm/backends/HIP/kernel/predict_kernel.hip.hpp
#	include/plssvm/backends/Kokkos/kernel/cg_explicit/blas.hpp
#	include/plssvm/backends/Kokkos/kernel/predict_kernel.hpp
#	include/plssvm/backends/OpenCL/kernel/cg_explicit/blas.cl
#	include/plssvm/backends/OpenCL/kernel/predict_kernel.cl
#	include/plssvm/backends/SYCL/kernel/predict/work_group/predict_kernel.hpp
#	include/plssvm/backends/gpu_device_ptr.hpp
#	include/plssvm/constants.hpp
#	include/plssvm/matrix.hpp
#	include/plssvm/model/classification_model.hpp
#	include/plssvm/svm/csvc.hpp
#	include/plssvm/svm/csvm.hpp
#	src/plssvm/backends/CUDA/detail/device_ptr.cu
#	src/plssvm/backends/HIP/detail/device_ptr.hip
#	src/plssvm/backends/Kokkos/csvm.cpp
#	src/plssvm/backends/Kokkos/detail/device_ptr.cpp
#	src/plssvm/backends/OpenCL/detail/device_ptr.cpp
#	src/plssvm/svm/csvm.cpp
#	tests/backends/generic_base_csvm_tests.hpp
#	tests/backends/generic_csvm_tests.hpp
#	tests/backends/generic_device_ptr_tests.hpp
#	tests/backends/generic_gpu_csvm_tests.hpp
#	tests/custom_test_macros.hpp
#	tests/data_set/classification/constructors.cpp
#	tests/data_set/regression/constructors.cpp
#	tests/detail/data_distribution.cpp
#	tests/matrix.cpp
#	tests/svm/csvc.cpp
#	tests/svm/csvr.cpp

+++ src/plssvm/backends/CUDA/detail/device_ptr.cu
@@ -15,6 +15,8 @@
 #include "plssvm/exceptions/exceptions.hpp"                    // plssvm::exception
 #include "plssvm/shape.hpp"                                    // plssvm::shape
 
+#include "driver_types.h"  // cudaMemcpyHostToDevice, cudaMemcpyDeviceToHost, cudaMemcpyDeviceToDevice
+
 #include "fmt/format.h"  // fmt::format
 
 #include <algorithm>  // std::min
@@ -35,7 +37,7 @@ device_ptr<T>::device_ptr(const plssvm::shape shape, const queue_type device) :
 template <typename T>
 device_ptr<T>::device_ptr(const plssvm::shape shape, const plssvm::shape padding, const queue_type device) :
     base_type{ shape, padding, device } {
-    if (queue_ < 0 || queue_ >= static_cast<int>(get_device_count())) {
+    if (queue_ < 0 || queue_ >= get_device_count()) {
         throw backend_exception{ fmt::format("Illegal device ID! Must be in range: [0, {}) but is {}.", get_device_count(), queue_) };
     }
     detail::set_device(queue_);
